trigger: Move animaton code to separate class.
