+++ lib/vpt/trigger/trigger-animation.ts
@@ -1,0 +1,104 @@
+/*
+ * VPDB - Virtual Pinball Database
+ * Copyright (C) 2019 freezy <freezy@vpdb.io>
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { Table } from '../..';
+import { IAnimation } from '../../game/ianimatable';
+import { Player } from '../../game/player';
+import { Trigger } from './trigger';
+import { TriggerData } from './trigger-data';
+import { TriggerEvents } from './trigger-events';
+import { TriggerState } from './trigger-state';
+
+export class TriggerAnimation implements IAnimation {
+
+	private readonly data: TriggerData;
+	private readonly state: TriggerState;
+	private readonly events: TriggerEvents;
+
+	private timeMsec = 0;
+	private doAnimation: boolean = false;
+	private moveDown: boolean = false;
+
+	constructor(data: TriggerData, state: TriggerState, events: TriggerEvents) {
+		this.data = data;
+		this.state = state;
+		this.events = events;
+	}
+
+	public init(player: Player): void {
+		// nothing to init.
+	}
+
+	public updateAnimation(player: Player, table: Table) {
+		const oldTimeMsec = (this.timeMsec < player.timeMsec) ? this.timeMsec : player.timeMsec;
+		this.timeMsec = player.timeMsec;
+		const diffTimeMsec = player.timeMsec - oldTimeMsec;
+
+		let animLimit = this.data.shape === Trigger.ShapeTriggerStar ? this.data.radius * (1.0 / 5.0) : 32.0;
+		if (this.data.shape === Trigger.ShapeTriggerButton) {
+			animLimit = this.data.radius * (1.0 / 10.0);
+		}
+		if (this.data.shape === Trigger.ShapeTriggerWireC) {
+			animLimit = 60.0;
+		}
+		if (this.data.shape === Trigger.ShapeTriggerWireD) {
+			animLimit = 25.0;
+		}
+
+		const limit = animLimit * table.getScaleZ();
+
+		if (this.events.hitEvent) {
+			this.doAnimation = true;
+			this.events.hitEvent = false;
+			// unhitEvent = false;   // Bugfix: If HitEvent and unhitEvent happen at the same time, you want to favor the unhit, otherwise the switch gets stuck down.
+			this.state.heightOffset = 0.0;
+			this.moveDown = true;
+		}
+		if (this.events.unhitEvent) {
+			this.doAnimation = true;
+			this.events.unhitEvent = false;
+			this.events.hitEvent = false;
+			this.state.heightOffset = limit;
+			this.moveDown = false;
+		}
+
+		if (this.doAnimation) {
+			let step = diffTimeMsec * this.data.animSpeed * table.getScaleZ();
+			if (this.moveDown) {
+				step = -step;
+			}
+			this.state.heightOffset += step;
+
+			if (this.moveDown) {
+				if (this.state.heightOffset <= -limit) {
+					this.state.heightOffset = -limit;
+					this.doAnimation = false;
+					this.moveDown = false;
+				}
+
+			} else {
+				if (this.state.heightOffset >= 0.0) {
+					this.state.heightOffset = 0.0;
+					this.doAnimation = false;
+					this.moveDown = true;
+				}
+			}
+		}
+	}
+}
