+++ lib/vpt/trigger/trigger-events.ts
@@ -17,35 +17,23 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-import { Table } from '../..';
-import { IAnimation } from '../../game/ianimatable';
 import { IPlayable } from '../../game/iplayable';
-import { Player } from '../../game/player';
 import { FireEvents } from '../../physics/fire-events';
-import { Trigger } from './trigger';
 import { TriggerData } from './trigger-data';
 import { TriggerState } from './trigger-state';
 
-export class TriggerEvents extends FireEvents implements IAnimation {
+export class TriggerEvents extends FireEvents {
 
 	private readonly data: TriggerData;
 	private readonly state: TriggerState;
 
-	private hitEvent = false;
-	private unhitEvent = false;
-
-	private timeMsec = 0;
-	private doAnimation: boolean = false;
-	private moveDown: boolean = false;
+	public hitEvent = false;
+	public unhitEvent = false;
 
 	constructor(data: TriggerData, state: TriggerState, playable: IPlayable) {
 		super(playable);
 		this.data = data;
 		this.state = state;
-	}
-
-	public init(player: Player): void {
-		// nothing to init.
 	}
 
 	public triggerAnimationHit(): void {
@@ -54,62 +42,5 @@ export class TriggerEvents extends FireEvents implements IAnimation {
 
 	public triggerAnimationUnhit(): void {
 		this.unhitEvent = true;
-	}
-
-	public updateAnimation(player: Player, table: Table) {
-		const oldTimeMsec = (this.timeMsec < player.timeMsec) ? this.timeMsec : player.timeMsec;
-		this.timeMsec = player.timeMsec;
-		const diffTimeMsec = player.timeMsec - oldTimeMsec;
-
-		let animLimit = this.data.shape === Trigger.ShapeTriggerStar ? this.data.radius * (1.0 / 5.0) : 32.0;
-		if (this.data.shape === Trigger.ShapeTriggerButton) {
-			animLimit = this.data.radius * (1.0 / 10.0);
-		}
-		if (this.data.shape === Trigger.ShapeTriggerWireC) {
-			animLimit = 60.0;
-		}
-		if (this.data.shape === Trigger.ShapeTriggerWireD) {
-			animLimit = 25.0;
-		}
-
-		const limit = animLimit * table.getScaleZ();
-
-		if (this.hitEvent) {
-			this.doAnimation = true;
-			this.hitEvent = false;
-			// unhitEvent = false;   // Bugfix: If HitEvent and unhitEvent happen at the same time, you want to favor the unhit, otherwise the switch gets stuck down.
-			this.state.heightOffset = 0.0;
-			this.moveDown = true;
-		}
-		if (this.unhitEvent) {
-			this.doAnimation = true;
-			this.unhitEvent = false;
-			this.hitEvent = false;
-			this.state.heightOffset = limit;
-			this.moveDown = false;
-		}
-
-		if (this.doAnimation) {
-			let step = diffTimeMsec * this.data.animSpeed * table.getScaleZ();
-			if (this.moveDown) {
-				step = -step;
-			}
-			this.state.heightOffset += step;
-
-			if (this.moveDown) {
-				if (this.state.heightOffset <= -limit) {
-					this.state.heightOffset = -limit;
-					this.doAnimation = false;
-					this.moveDown = false;
-				}
-
-			} else {
-				if (this.state.heightOffset >= 0.0) {
-					this.state.heightOffset = 0.0;
-					this.doAnimation = false;
-					this.moveDown = true;
-				}
-			}
-		}
 	}
 }
